+++ src/Http/Controller/TenantController.php
@@ -1,9 +1,19 @@
 <?php
-
 /*
- * This file is part of the current application.
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- * (c) Dave Redfern
+ * This software consists of voluntary contributions made by many individuals
+ * and is licensed under the MIT license.
  */
 
 namespace Somnambulist\Tenancy\Http\Controller;
@@ -21,8 +31,8 @@ use Illuminate\Http\Request;
  *
  * Note: this controller should be wrapped in the Auth middleware.
  *
- * @package    App\Http\Controllers
- * @subpackage App\Http\Controllers\TenantController
+ * @package    Somnambulist\Tenancy\Http\Controller
+ * @subpackage Somnambulist\Tenancy\Http\Controller\TenantController
  * @author     Dave Redfern
  */
 abstract class TenantController extends Controller